Restructure backend tests
