+++ package.json
@@ -128,6 +128,7 @@
     "eslint-plugin-cypress": "^2.11.2",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-mocha": "^8.0.0",
     "eslint-plugin-react": "^7.21.1",
     "eslint-plugin-react-hooks": "3.0.0",
     "husky": "^3.0.5",
@@ -146,14 +147,15 @@
     "analyze": "cross-env ANALYZE=true yarn build",
     "sitemap": "cross-env node next-sitemap.js",
     "cypress": "cross-env cypress run",
+    "cy:open": "cross-env cypress open",
     "test:frontend": "NODE_ENV=TEST_FRONTEND NODE_OPTIONS=--max_old_space_size=4096 start-server-and-test 'yarn dev' 9000 'yarn cypress'",
+    "test:backend:mock": "NODE_ENV=TEST_FRONTEND SERVER_ONLY=true mocha --reporter spec --timeout 60000 --exit",
     "test:backend": "NODE_ENV=TEST_BACKEND SERVER_ONLY=true mocha --reporter spec --timeout 60000 --exit",
     "test": "npm run test:frontend && npm run test:backend",
     "cover:frontend": "NODE_ENV=test nyc --silent --no-clean npm run test:frontend",
     "cover:backend": "nyc --silent --no-clean npm run test:backend ",
     "cover:report": "nyc report --reporter=lcov --reporter=text",
-    "cover": "npm run cover:frontend && npm run cover:backend && npm run cover:report",
-    "cy:open": "cross-env cypress open"
+    "cover": "npm run cover:frontend && npm run cover:backend && npm run cover:report"
   },
   "engines": {
     "node": ">=14.15"
