+++ test/payload/user.js
@@ -1,15 +1,15 @@
 module.exports = {
-  "id":"59b01b255214e50001ce7f93x",
-  "_id":"59b01b585214e50001ce7f93x",
-  "email":"john@doe.com",
-  "name":"John Doe",
-  "photo":"https://s3.amazonaws.com/wri-api-backups/resourcewatch/staging/profiles/avatars/000/000/039/original/data?1549875449",
-  "createdAt":"2017-09-06T15:58:32.368Z",
-  "updatedAt":null,
-  "role":"ADMIN",
-  "provider":"local",
-  "extraUserData":{
-      "apps":["rw","gfw","prep","aqueduct","gfw-climate","earthhq"]
+  id: '59b01b255214e50001ce7f93x',
+  _id: '59b01b585214e50001ce7f93x',
+  email: 'john@doe.com',
+  name: 'John Doe',
+  photo: 'https://s3.amazonaws.com/wri-api-backups/resourcewatch/staging/profiles/avatars/000/000/039/original/data?1549875449',
+  createdAt: '2017-09-06T15:58:32.368Z',
+  updatedAt: null,
+  role: 'ADMIN',
+  provider: 'local',
+  extraUserData: {
+    apps: ['rw', 'gfw', 'prep', 'aqueduct', 'gfw-climate', 'earthhq'],
   },
-  "token":"valid_token"
+  token: 'valid_token',
 };
